requirements.txt
eth_ape==0.7.14
giza_actions==0.3.0
giza_cli==0.15.0
numpy==1.26.4
python-dotenv
starknet-py
pandas==1.5.3
prefect==2.14.6
torch
# python-dotenv==1.0.1
# scikit_learn==1.4.2
# torch==2.2.1
# torch==2.2.2

